Fix autoreload with embedded server (#2397)
* KTOR-2214 Fix autoreload in embeddedServer

* KTOR-2082 Fix flaky server start
