+++ ktor-client/ktor-client-tests/jvm/src/io/ktor/client/tests/utils/TestTcpServer.kt
@@ -8,10 +8,11 @@ import io.ktor.network.selector.*
 import io.ktor.network.sockets.*
 import io.ktor.network.sockets.Socket
 import kotlinx.coroutines.*
+import java.io.*
 import java.net.*
 import kotlin.coroutines.*
 
-internal class TestTcpServer(val port: Int, handler: suspend (Socket) -> Unit) : CoroutineScope {
+internal class TestTcpServer(val port: Int, handler: suspend (Socket) -> Unit) : CoroutineScope, Closeable {
     private val selector = ActorSelectorManager(Dispatchers.IO)
     override val coroutineContext: CoroutineContext
 
@@ -36,7 +37,5 @@ internal class TestTcpServer(val port: Int, handler: suspend (Socket) -> Unit) :
         }
     }
 
-    public fun close() {
-        coroutineContext.cancel()
-    }
+    override fun close() { coroutineContext.cancel() }
 }
